+++ src/wine-host/bridges/vst3.h
@@ -244,6 +244,14 @@ struct Vst3PluginInstance {
      * have, but we'll just do this out of precaution.
      */
     bool is_initialized = false;
+
+    /**
+     * Whether the plugin instance is currently in offline processing mode or
+     * not. Needed as a HACK for IK Multimedia's T-RackS 5 because those plugins
+     * will deadlock if they receive a timer proc on the Win32 message loop
+     * while doing offline processing.
+     */
+    bool is_offline_processing = false;
 };
 
 /**